# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Task Application",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/branch": {
            "post": {
                "tags": [
                    "Branch"
                ],
                "summary": "Create a new branch",
                "description": "Create a new branch.",
                "operationId": "d7046dc51dd911930648bea9c3009624",
                "requestBody": {
                    "description": "Branch data",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "brand_id",
                                    "region_id",
                                    "district_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "branch 1"
                                    },
                                    "images[]": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    },
                                    "brand_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "region_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "district_id": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Branch created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Branch created successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "name": {
                                                    "properties": {
                                                        "name": {
                                                            "type": "string"
                                                        },
                                                        "brand_id": {
                                                            "type": "integer"
                                                        },
                                                        "region_id": {
                                                            "type": "integer"
                                                        },
                                                        "district_id": {
                                                            "type": "integer"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/branches": {
            "get": {
                "tags": [
                    "Branch"
                ],
                "summary": "Get all branches",
                "description": "Get a list of all branches.",
                "operationId": "648e90f3a9ff3b032011f59b611e719a",
                "responses": {
                    "200": {
                        "description": "List of branches",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "name": {
                                                "type": "string"
                                            },
                                            "brand_id": {
                                                "type": "integer"
                                            },
                                            "region_id": {
                                                "type": "integer"
                                            },
                                            "district_id": {
                                                "type": "integer"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/branch/{id}": {
            "post": {
                "tags": [
                    "Branch"
                ],
                "summary": "Update branch details",
                "description": "Update a branch's details.",
                "operationId": "f1747aacfbbdf5d112d7ef6600fe49f6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Branch ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Branch data",
                    "required": false,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "branch 1"
                                    },
                                    "images[]": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    },
                                    "brand_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "region_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "district_id": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Branch created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Branch created successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "name": {
                                                    "properties": {
                                                        "name": {
                                                            "type": "string"
                                                        },
                                                        "brand_id": {
                                                            "type": "integer"
                                                        },
                                                        "region_id": {
                                                            "type": "integer"
                                                        },
                                                        "district_id": {
                                                            "type": "integer"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Branch not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Branch not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Branch"
                ],
                "summary": "Delete branch by ID",
                "description": "Delete a branch by ID.",
                "operationId": "94918f1dc2800109c7c36b4ec507ab6f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the branch to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Branch deleted successfully"
                    },
                    "404": {
                        "description": "Branch not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/branch/{region_id}": {
            "get": {
                "tags": [
                    "Branch"
                ],
                "summary": "Get branch counts by region",
                "operationId": "5dcc0e46b357716441c04a12252f171d",
                "parameters": [
                    {
                        "name": "region_id",
                        "in": "path",
                        "description": "ID of the region",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "region": {
                                            "description": "Name of the region",
                                            "type": "string"
                                        },
                                        "districts": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "name": {
                                                        "description": "Name of the district",
                                                        "type": "string"
                                                    },
                                                    "brands": {
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "name": {
                                                                    "description": "Name of the brand",
                                                                    "type": "string"
                                                                },
                                                                "countOfBranches": {
                                                                    "description": "Number of branches for the brand in the district",
                                                                    "type": "integer"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Region not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}